Use new CM service-worker asset
See https://github.com/cargomedia/cm/pull/2177
